refactor: improve paddings
