+++ ui/src/main/java/com/example/ui/screens/searchByActor/SearchByActorScreen.kt
@@ -1,6 +1,7 @@
 package com.example.ui.screens.searchByActor
 
 
+import android.graphics.Movie
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
@@ -13,14 +14,14 @@ import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.navigation.NavController
 import com.example.designsystem.R
-import com.example.designsystem.components.LoadingIndicator
 import com.example.designsystem.components.MovieCard
 import com.example.designsystem.components.NoDataContainer
 import com.example.designsystem.components.TextField
@@ -28,75 +29,82 @@ import com.example.designsystem.components.appBar.DefaultAppBar
 import com.example.designsystem.theme.AflamiTheme
 import com.example.designsystem.utils.ThemeAndLocalePreviews
 import com.example.ui.application.LocalNavController
-import com.example.viewmodel.searchByActor.SearchByActorUiState
+import com.example.viewmodel.search.countrySearch.MovieUiState
+import com.example.viewmodel.searchByActor.SearchByActorEffect
 import com.example.viewmodel.searchByActor.SearchByActorViewModel
 import org.koin.androidx.compose.koinViewModel
 
 @Composable
 fun SearchByActorScreen(
     modifier: Modifier = Modifier,
-    viewModel: SearchByActorViewModel = koinViewModel()
-) {
-    val uiState by viewModel.state.collectAsStateWithLifecycle()
+    viewModel: SearchByActorViewModel=koinViewModel(),
+){
+    val uiState = viewModel.state.collectAsStateWithLifecycle()
     val navController = LocalNavController.current
-    SearchByActorContent(
-        modifier = modifier,
-        onNavigateBackClicked = {
-            navController.popBackStack()
-        },
-        onValueChange = {
-            viewModel.onQueryChange(it)
-        },
-        state = uiState
+    LaunchedEffect(Unit) {
+        viewModel.effect.collect { effect ->
+            when (effect) {
+
+                SearchByActorEffect.NavigateBack -> {
+                    navController.popBackStack()
+                }
+
+                SearchByActorEffect.NoInternetConnection -> {}
+            }
+        }
+    }
+    SearchByActorContent (
+        modifier=modifier,
+        onNavigateBackClicked = {viewModel.onBackClicked()},
+        onValueChange = {viewModel.onQueryChange(it)}
     )
+
+    uiState.value.movies
+
 }
 
 @Composable
 private fun SearchByActorContent(
     modifier: Modifier = Modifier,
-    onNavigateBackClicked: () -> Unit,
-    onValueChange: (String) -> Unit,
-    state: SearchByActorUiState
+    onNavigateBackClicked: () -> Unit ,
+    onValueChange: (String) -> Unit ,
+    result: List<MovieUiState> = emptyList()
 ) {
+        Column(
+            modifier = modifier
+                .fillMaxSize()
+                .statusBarsPadding()
+                .padding(horizontal = 16.dp)
+        ) {
+            DefaultAppBar(
+                title = stringResource(R.string.find_by_actor),
+                showNavigateBackButton = true,
+                onNavigateBackClicked = {onNavigateBackClicked()}
+            )
+            TextField(
+                text="",
+                hintText = stringResource(R.string.find_by_actor),
+                onValueChange = {onValueChange(it)},
 
-    Column(
-        modifier = modifier
-            .statusBarsPadding()
-            .fillMaxSize()
-            .padding(horizontal = 16.dp)
-    ) {
-        DefaultAppBar(
-            title = stringResource(R.string.find_by_actor),
-            showNavigateBackButton = true,
-            onNavigateBackClicked = { onNavigateBackClicked() }
-        )
-        TextField(
-            text = "",
-            hintText = "Tom hanks",
-            onValueChange = { onValueChange(it) },
-        )
-        if (state.isLoading) {
-            Spacer(modifier = Modifier.weight(1f))
-            LoadingIndicator()
-            Spacer(modifier = Modifier.weight(1f))
-        } else {
-            if (state.movies.isEmpty()) {
+            )
+            if (result.isEmpty()) {
                 NoDataContainer(
                     imageRes = painterResource(R.drawable.placeholder_no_result_found),
                     title = stringResource(R.string.no_search_result),
-                    description = stringResource(R.string.no_search_result_description)
+                    description = stringResource(R.string.no_search_result_description),
                 )
             } else {
                 LazyVerticalGrid(
+                    modifier = Modifier.padding(horizontal = 16.dp),
                     columns = GridCells.Fixed(2),
                     contentPadding = PaddingValues(vertical = 12.dp),
                     horizontalArrangement = Arrangement.spacedBy(8.dp),
                     verticalArrangement = Arrangement.spacedBy(8.dp),
                 ) {
-                    items(state.movies) { movie ->
+                    items(result) {movie->
                         MovieCard(
                             movieImage = movie.poster,
-                            movieType = stringResource(R.string.movies),
+                            movieType = "Movies",
                             movieYear = movie.productionYear.toString(),
                             movieTitle = movie.name,
                             movieRating = movie.rating.toString(),
@@ -109,13 +117,15 @@ private fun SearchByActorContent(
             }
         }
     }
-}
 
 
 @Composable
 @ThemeAndLocalePreviews
 private fun SearchByActorContentPreview() {
     AflamiTheme {
-        SearchByActorScreen()
+        SearchByActorContent(
+            onNavigateBackClicked = {},
+            onValueChange = {}
+        )
     }
 }
